Use test_config file to setup a test blockchain for integration tests
