+++ test_config/test_config.go
@@ -20,6 +20,12 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/ethereum/go-ethereum/rpc"
+	"github.com/makerdao/vulcanizedb/pkg/eth"
+	"github.com/makerdao/vulcanizedb/pkg/eth/client"
+	"github.com/makerdao/vulcanizedb/pkg/eth/converters"
+	"github.com/makerdao/vulcanizedb/pkg/eth/node"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 
@@ -71,6 +77,41 @@ func SetTestConfig() {
 func setABIPath() {
 	gp := os.Getenv("GOPATH")
 	ABIFilePath = gp + "/src/github.com/makerdao/vulcanizedb/pkg/eth/testing/"
+}
+
+func NewTestBlockchain() (core.BlockChain, error) {
+	ipc := TestClient.IPCPath
+	// If we don't have an ipc path in the config file, check the env variable
+	if ipc == "" {
+		configErr := viper.BindEnv("url", "CLIENT_IPCPATH")
+		if configErr != nil {
+			return nil, fmt.Errorf("unable to bind url to CLIENT_IPCPATH env var %w", configErr)
+		}
+		ipc = viper.GetString("url")
+	}
+
+	rpcClient, ethClient, clientErr := getClients(ipc)
+	if clientErr != nil {
+		return nil, fmt.Errorf("failed to get test clients: %w", clientErr)
+	}
+
+	return getBlockChain(rpcClient, ethClient)
+}
+
+func getClients(ipc string) (client.RpcClient, *ethclient.Client, error) {
+	raw, err := rpc.Dial(ipc)
+	if err != nil {
+		return client.RpcClient{}, &ethclient.Client{}, err
+	}
+	return client.NewRpcClient(raw, ipc), ethclient.NewClient(raw), nil
+}
+
+func getBlockChain(rpcClient client.RpcClient, ethClient *ethclient.Client) (core.BlockChain, error) {
+	testClient := client.NewEthClient(ethClient)
+	testNode := node.MakeNode(rpcClient)
+	transactionConverter := converters.NewTransactionConverter(testClient)
+	blockChain := eth.NewBlockChain(testClient, rpcClient, testNode, transactionConverter)
+	return blockChain, nil
 }
 
 func NewTestDB(node core.Node) *postgres.DB {
